pyproject.toml
[project]
name = "postgres-py-issue-550"
version = "0.1.0"
description = "Postgrest query builder not being immutable test"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.1",
    "supabase>=2.10.0",
]
